SAVE-FILE-TOBUCKET/requirements.txt
requests
google-cloud-storage==1.41.0